fastapi/requirements.txt
anyio==3.6.2
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
fastapi==0.88.0
h11==0.14.0
idna==3.4
mongoengine==0.24.2
pydantic==1.10.2
pyfiglet==0.8.post1
pymongo==4.3.3
python-dotenv==0.21.0
sniffio==1.3.0
starlette==0.22.0
typing-extensions==4.4.0
uvicorn==0.20.0
